[Discover][Oblt] Move `getDocViewer` into the main observability root profile (#225610)
close https://github.com/elastic/kibana/issues/225408

Follow on
https://github.com/elastic/kibana/pull/224894#issuecomment-3004090408

Since the Attributes tab is now enabled by default, the separate
observability-root-profile-with-attributes-tab profile is no longer
needed. Moved getDocViewer into the main observability root profile to
simplify the code.

Functionality remains unchanged; this is a code cleanup.

(cherry picked from commit 55b1ef515f12f5bc9516143ce51ec07e4ad679dd)
