+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/observability/observability_root_profile/index.ts
@@ -7,17 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import type { ProfileProviderServices } from '../../profile_provider_services';
-import { createObservabilityRootProfileProvider } from './profile';
-import { createObservabilityRootProfileProviderWithAttributesTab } from './sub_profiles/observability_root_profile_with_attributes_tab';
-
-export const createObservabilityRootProfileProviders = (
-  providerServices: ProfileProviderServices
-) => {
-  const observabilityRootProfileProvider = createObservabilityRootProfileProvider(providerServices);
-
-  return [
-    createObservabilityRootProfileProviderWithAttributesTab(observabilityRootProfileProvider),
-    observabilityRootProfileProvider,
-  ];
-};
+export { createObservabilityRootProfileProvider } from './profile';
