+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/observability/observability_root_profile/profile.test.ts
@@ -10,6 +10,8 @@
 import { SolutionType } from '../../../profiles';
 import { createContextAwarenessMocks } from '../../../__mocks__';
 import { createObservabilityRootProfileProvider } from './profile';
+import { buildDataTableRecord } from '@kbn/discover-utils';
+import { DocViewsRegistry } from '@kbn/unified-doc-viewer';
 
 const mockServices = createContextAwarenessMocks().profileProviderServices;
 
@@ -90,4 +92,157 @@ describe('observabilityRootProfileProvider', () => {
       expect(defaultDataViews).toEqual([]);
     });
   });
+
+  describe('getDocViewer', () => {
+    it('does NOT add attributes doc viewer tab to the registry when the record has no attributes fields', () => {
+      const getDocViewer = observabilityRootProfileProvider.profile.getDocViewer!(
+        () => ({
+          title: 'test title',
+          docViewsRegistry: (registry) => registry,
+        }),
+        {
+          context: {
+            solutionType: SolutionType.Observability,
+            allLogsIndexPattern: mockServices.logsContextService.getAllLogsIndexPattern(),
+          },
+        }
+      );
+
+      const docViewer = getDocViewer({
+        record: buildMockRecord('test-index', {
+          foo: 'bar',
+        }),
+      });
+
+      const registry = new DocViewsRegistry();
+
+      expect(docViewer.title).toBe('test title');
+      expect(registry.getAll()).toHaveLength(0);
+
+      docViewer.docViewsRegistry(registry);
+
+      expect(registry.getAll()).toHaveLength(0);
+    });
+    it('adds attributes doc viwer tab to the registry when the record has any attributes. field', () => {
+      const getDocViewer = observabilityRootProfileProvider.profile.getDocViewer!(
+        () => ({
+          title: 'test title',
+          docViewsRegistry: (registry) => registry,
+        }),
+        {
+          context: {
+            solutionType: SolutionType.Observability,
+            allLogsIndexPattern: mockServices.logsContextService.getAllLogsIndexPattern(),
+          },
+        }
+      );
+
+      const docViewer = getDocViewer({
+        record: buildMockRecord('test-index', {
+          'attributes.foo': 'bar',
+        }),
+      });
+
+      const registry = new DocViewsRegistry();
+
+      expect(docViewer.title).toBe('test title');
+      expect(registry.getAll()).toHaveLength(0);
+      docViewer.docViewsRegistry(registry);
+
+      expect(registry.getAll()).toHaveLength(1);
+
+      expect(registry.getAll()[0]).toEqual(
+        expect.objectContaining({
+          id: 'doc_view_obs_attributes_overview',
+          title: 'Attributes',
+          order: 9,
+          component: expect.any(Function),
+        })
+      );
+    });
+    it('adds attributes doc viwer tab to the registry when the record has any scope.attributes. field', () => {
+      const getDocViewer = observabilityRootProfileProvider.profile.getDocViewer!(
+        () => ({
+          title: 'test title',
+          docViewsRegistry: (registry) => registry,
+        }),
+        {
+          context: {
+            solutionType: SolutionType.Observability,
+            allLogsIndexPattern: mockServices.logsContextService.getAllLogsIndexPattern(),
+          },
+        }
+      );
+
+      const docViewer = getDocViewer({
+        record: buildMockRecord('test-index', {
+          'scope.attributes.foo': 'bar',
+        }),
+      });
+
+      const registry = new DocViewsRegistry();
+
+      expect(docViewer.title).toBe('test title');
+      expect(registry.getAll()).toHaveLength(0);
+      docViewer.docViewsRegistry(registry);
+
+      expect(registry.getAll()).toHaveLength(1);
+
+      expect(registry.getAll()[0]).toEqual(
+        expect.objectContaining({
+          id: 'doc_view_obs_attributes_overview',
+          title: 'Attributes',
+          order: 9,
+          component: expect.any(Function),
+        })
+      );
+    });
+    it('adds attributes doc viewer tab to the registry when the record has any resource.attributes. field', () => {
+      const getDocViewer = observabilityRootProfileProvider.profile.getDocViewer!(
+        () => ({
+          title: 'test title',
+          docViewsRegistry: (registry) => registry,
+        }),
+        {
+          context: {
+            solutionType: SolutionType.Observability,
+            allLogsIndexPattern: mockServices.logsContextService.getAllLogsIndexPattern(),
+          },
+        }
+      );
+
+      const docViewer = getDocViewer({
+        record: buildMockRecord('test-index', {
+          'resource.attributes.foo': 'bar',
+        }),
+      });
+
+      const registry = new DocViewsRegistry();
+
+      expect(docViewer.title).toBe('test title');
+      expect(registry.getAll()).toHaveLength(0);
+      docViewer.docViewsRegistry(registry);
+
+      expect(registry.getAll()).toHaveLength(1);
+
+      expect(registry.getAll()[0]).toEqual(
+        expect.objectContaining({
+          id: 'doc_view_obs_attributes_overview',
+          title: 'Attributes',
+          order: 9,
+          component: expect.any(Function),
+        })
+      );
+    });
+  });
 });
+
+const buildMockRecord = (index: string, fields: Record<string, unknown> = {}) =>
+  buildDataTableRecord({
+    _id: '',
+    _index: index,
+    fields: {
+      _index: index,
+      ...fields,
+    },
+  });
