+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/observability/observability_root_profile/profile.tsx
@@ -10,7 +10,7 @@
 import { SolutionType } from '../../../profiles';
 import type { ProfileProviderServices } from '../../profile_provider_services';
 import { OBSERVABILITY_ROOT_PROFILE_ID } from '../consts';
-import { createGetAppMenu, getDefaultAdHocDataViews } from './accessors';
+import { createGetAppMenu, getDefaultAdHocDataViews, getDocViewer } from './accessors';
 import type { ObservabilityRootProfileProvider } from './types';
 
 export const createObservabilityRootProfileProvider = (
@@ -20,6 +20,7 @@ export const createObservabilityRootProfileProvider = (
   profile: {
     getAppMenu: createGetAppMenu(services),
     getDefaultAdHocDataViews,
+    getDocViewer,
   },
   resolve: (params) => {
     if (params.solutionNavId !== SolutionType.Observability) {
